requirements.txt
falcon
rethinkdb
pyyaml